requirements.txt
opencv-python
Pillow
tqdm
scikit-image